Fix crash when showing recovery view
The ListBox was filled with instances of the Item class defined in
data_recovery_view.cpp but for some reason the app called the methods of
the ButtonSet::Item class instead, leading to a crash in
ButtonSet::Item::onPaint() when trying to access m_icon.

Renaming the Item class to a more unique name fixes this.